Fix web clipper behaviour for tabs
Grouping of similar webpages do not work for tabs-as-list since `pageUrl` is
set to `about:blank`, so all lists of tabs end up in the same article.

This fix skips the grouping if page is `about:blank` so the list of tabs is
stored as a new article by default.